requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
black==24.10.0
cattrs==24.1.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.3
decorator==5.1.1
dj-database-url==2.2.0
Django==5.1.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-login-required-middleware==0.9.0
django-phonenumber-field==8.0.0
django-schema-graph==3.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
executing==2.1.0
Faker==30.0.0
idna==3.10
ipython==8.27.0
jedi==0.19.1
kombu==5.4.2
Markdown==3.7
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
phonenumberslite==8.13.47
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg==3.2.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.1
requests==2.32.3
ruff==0.7.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
